Modify the code to leverage the Common Integrations and Module integration
