+++ src/Events/Integrations/Integration_Abstract.php
@@ -2,6 +2,8 @@
 
 namespace TEC\Events\Integrations;
 
+use TEC\Common\Integrations\Integration_Abstract as Common_Integration_Abstract;
+
 /**
  * Class Integration_Abstract
  *
@@ -9,54 +11,20 @@ namespace TEC\Events\Integrations;
  *
  * @package TEC\Events\Integrations
  */
-abstract class Integration_Abstract extends \tad_DI52_ServiceProvider {
-	/**
-	 * Binds and sets up implementations.
-	 *
-	 * @since 6.0.4
-	 */
-	public function register() {
-		// Registers this provider as a singleton for ease of use.
-		$this->container->singleton( self::class, self::class );
-
-		// Prevents any loading in case we shouldn't load.
-		if ( ! $this->should_load() ) {
-			return;
-		}
-
-		$this->load();
-	}
-
-	/**
-	 * Gets the slug for this integration.
-	 *
-	 * @since 6.0.4
-	 *
-	 * @return string
-	 */
-	abstract public static function get_slug(): string;
-
-	/**
-	 * Determines whether this integration should load.
-	 *
-	 * @since 6.0.4
-	 *
-	 * @return bool
-	 */
-	public function should_load(): bool {
-		return $this->filter_should_load( $this->load_conditionals() );
-	}
-
+abstract class Integration_Abstract extends Common_Integration_Abstract {
 	/**
 	 * Filters whether the integration should load.
 	 *
 	 * @since 6.0.4
+	 * @since TBD uses the Common integration as the base filter and then Events for Legacy compatibility.
 	 *
 	 * @param bool $value Whether the integration should load.
 	 *
 	 * @return bool
 	 */
 	protected function filter_should_load( bool $value ): bool {
+		$value = parent::filter_should_load( $value );
+
 		$slug = static::get_slug();
 		$type = static::get_type();
 
@@ -90,31 +58,4 @@ abstract class Integration_Abstract extends \tad_DI52_ServiceProvider {
 		 */
 		return (bool) apply_filters( "tec_events_integrations_{$type}_{$slug}_should_load", $value );
 	}
-
-	/**
-	 * Determines if the integration in question should be loaded.
-	 *
-	 * @since 6.0.4
-	 *
-	 * @return bool
-	 */
-	abstract public function load_conditionals(): bool;
-
-	/**
-	 * Loads the integration itself.
-	 *
-	 * @since 6.0.4
-	 *
-	 * @return void
-	 */
-	abstract protected function load(): void;
-
-	/**
-	 * Determines the integration type.
-	 *
-	 * @since 6.0.4
-	 *
-	 * @return string
-	 */
-	abstract public static function get_type(): string;
 }
