+++ src/Events/Integrations/Modules/Ticket_Emails/Email/Ticket.php
@@ -1,47 +1,48 @@
 <?php
 /**
- * Class RSVP.
+ * Class Ticket.
  *
  * @since TBD
  *
- * @package TEC\Events\Integrations\Plugins\Event_Tickets
+ * @package TEC\Events\Integrations\Modules\Ticket_Emails
  */
 
-namespace TEC\Events\Integrations\Plugins\Event_Tickets\Emails\Email;
+namespace TEC\Events\Integrations\Modules\Ticket_Emails\Email;
 
-use TEC\Tickets\Emails\Email\RSVP as RSVP_Email;
-use TEC\Events\Integrations\Plugins\Event_Tickets\Emails\Emails as TEC_Email_Handler;
-use TEC\Events\Integrations\Plugins\Event_Tickets\Emails\Template as Template;
+use TEC\Events\Integrations\Modules\Ticket_Emails\Emails as TEC_Email_Handler;
+use TEC\Events\Integrations\Modules\Ticket_Emails\Template as Template;
+use TEC\Tickets\Emails\Email\Ticket as Ticket_Email;
+use TEC\Tickets\Emails\Email_Abstract;
 
 /**
- * Class RSVP.
+ * Class Ticket.
  *
  * @since TBD
  *
  * @package TEC\Events\Integrations\Plugins\Event_Tickets
  */
-class RSVP {
+class Ticket {
 	/**
 	 * The option key for the Event calendar links.
 	 *
-	 * @see TEC\Tickets\Emails\Email_Abstract::get_option_key() for option key format.
+	 * @see Email_Abstract::get_option_key() for option key format.
 	 *
 	 * @since TBD
 	 *
 	 * @var string
 	 */
-	public static $option_add_event_links = 'tec-tickets-emails-rsvp-add-event-links';
+	public static $option_add_event_links = 'tec-tickets-emails-ticket-add-event-links';
 
 	/**
 	 * The option key for the Event calendar invite.
 	 *
-	 * @see TEC\Tickets\Emails\Email_Abstract::get_option_key() for option key format.
+	 * @see Email_Abstract::get_option_key() for option key format.
 	 *
 	 * @since TBD
 	 *
 	 * @var string
 	 */
-	public static $option_add_event_ics = 'tec-tickets-emails-rsvp-add-event-ics';
+	public static $option_add_event_ics = 'tec-tickets-emails-ticket-add-event-ics';
 
 	/**
 	 * Filter the email settings and add TEC specific settings.
@@ -52,7 +53,7 @@ class RSVP {
 	 *
 	 * @return array<array<string,mixed>> $settings The modified email settings.
 	 */
-	public function filter_tec_tickets_emails_rsvp_email_settings( $settings ) {
+	public function filter_tec_tickets_emails_ticket_email_settings( $settings ) {
 
 		$settings[ self::$option_add_event_links ] = [
 			'type'            => 'toggle',
@@ -65,7 +66,7 @@ class RSVP {
 		$settings[ self::$option_add_event_ics ] = [
 			'type'            => 'toggle',
 			'label'           => esc_html__( 'Attach Calendar Invites', 'the-events-calendar' ),
-			'tooltip'         => esc_html__( 'Attach calendar invites (.ics) to the RSVP email.', 'the-events-calendar' ),
+			'tooltip'         => esc_html__( 'Attach calendar invites (.ics) to the ticket email.', 'the-events-calendar' ),
 			'default'         => true,
 			'validation_type' => 'boolean',
 		];
@@ -74,29 +75,19 @@ class RSVP {
 	}
 
 	/**
-	 * Filters the attachments for the RSVP Emails and maybe add the calendar ics file.
+	 * Filters the attachments for the Tickets Emails and maybe add the calendar ics file.
 	 *
 	 * @since TBD
 	 *
-	 * @param array<string,string> $attachments  The attachments for the Tickets Emails.
+	 * @param array<string,string> $attachments The placeholders for the Tickets Emails.
 	 * @param string               $email_id     The email ID.
 	 * @param Email_Abstract       $email_class  The email class.
 	 *
-	 * @return array<string,string> The filtered attachments for the RSVP Emails.
+	 * @return array<string,string> The filtered attachments for the Tickets Emails.
 	 */
-	public function filter_tec_tickets_emails_rsvp_email_attachments( $attachments, $email_id, $email_class ) {
+	public function filter_tec_tickets_emails_ticket_email_attachments( $attachments, $email_id, $email_class ) {
 		if ( ! $email_class->is_enabled() ) {
 			return $attachments;
-		}
-
-		$use_ticket_email = tribe_get_option( $email_class->get_option_key( 'use-ticket-email' ), false );
-
-		if ( ! empty( $use_ticket_email ) ) {
-			$email_class = tribe( TEC\Tickets\Emails\Email\Ticket::class );
-
-			if ( ! $email_class->is_enabled() ) {
-				return $attachments;
-			}
 		}
 
 		if ( ! tribe_is_truthy( tribe_get_option( self::$option_add_event_ics, true ) ) ) {
@@ -121,7 +112,7 @@ class RSVP {
 	}
 
 	/**
-	 * Maybe include event links.
+	 * Maybe include the event links.
 	 *
 	 * @since TBD
 	 *
@@ -129,21 +120,11 @@ class RSVP {
 	 * @return void
 	 */
 	public function maybe_include_event_links( $et_template ) {
-		$email_class = tribe( RSVP_Email::class );
+		$email_class = tribe( Ticket_Email::class );
 
 		// Bail early if the email class is not enabled.
 		if ( ! $email_class->is_enabled() ) {
 			return;
-		}
-
-		$use_ticket_email = tribe_get_option( $email_class->get_option_key( 'use-ticket-email' ), false );
-
-		if ( ! empty( $use_ticket_email ) ) {
-			$email_class = tribe( TEC\Tickets\Emails\Email\Ticket::class );
-
-			if ( ! $email_class->is_enabled() ) {
-				return;
-			}
 		}
 
 		if ( ! tribe_is_truthy( tribe_get_option( self::$option_add_event_links, true ) ) ) {
