+++ src/Events/Integrations/Modules/Ticket_Emails/Emails.php
@@ -4,11 +4,12 @@
  *
  * @since   TBD
  *
- * @package TEC\Events\Integrations\Plugins\Event_Tickets
+ * @package TEC\Events\Integrations\Modules\Ticket_Emails
  */
 
-namespace TEC\Events\Integrations\Plugins\Event_Tickets\Emails;
+namespace TEC\Events\Integrations\Modules\Ticket_Emails;
 
+use TEC\Tickets\Emails\Email_Abstract;
 use Tribe__Events__Main as Main;
 use Tribe\Utils\Lazy_String;
 
@@ -17,7 +18,7 @@ use Tribe\Utils\Lazy_String;
  *
  * @since   TBD
  *
- * @package TEC\Events\Integrations\Plugins\Event_Tickets
+ * @package TEC\Events\Integrations\Modules\Ticket_Emails
  */
 class Emails {
 	/**
