+++ src/Events/Integrations/Modules/Ticket_Emails/Template.php
@@ -1,16 +1,16 @@
 <?php
 
-namespace TEC\Events\Integrations\Plugins\Event_Tickets\Emails;
+namespace TEC\Events\Integrations\Modules\Ticket_Emails;
 
-use \Tribe__Template;
 use Tribe__Events__Main as TEC;
+use Tribe__Template;
 
 /**
  * Class Template
  *
  * @since TBD
  *
- * @package TEC\Events\Integrations\Plugins\Event_Tickets
+ * @package TEC\Events\Integrations\Modules\Ticket_Emails
  */
 class Template extends Tribe__Template {
 	/**
