--- src/Events/Integrations/Themes/Theme_Integration.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace TEC\Events\Integrations\Themes;
-
-trait Theme_Integration {
-	/**
-	 * Gets the integration type.
-	 *
-	 * @since 6.0.4
-	 *
-	 * @return string
-	 */
-	public static function get_type(): string {
-		return 'theme';
-	}
-}
